Move validation scripts
