+++ .gitignore
@@ -1,4 +1,3 @@
 .DS_Store
-tests/node_modules
 node_modules/
 build
