+++ scripts/validation/metadata-schema.json
