+++ scripts/validation/package.json
