+++ scripts/validation/validator.js
